+++ CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added comments and help blocks to various functions
+- Added `Submit-FTPUpload` as public function
+- Modified `Get-ADHostAudit` to fix multiple string output
+
+### Changed
+
+- Updated documentation to include examples and usage information
+
+### Fixed
+
+- Fixed error handling in `Get-ADHostAudit` for blank output
+- Fixed several bugs related to incorrect variable naming and parameter types
+
+
+## [0.1.6] - 2023-02-20
+
+### Added
+
 - Added comment help block to `Get-ADUserLogonAudit`
 ## [0.1.5] - 2023-02-20